webinar_part_2/1.0_nxos_existing_no_netbox/requirements.txt
pynetbox
pyyaml
netaddr
ansible-vault
